add script to send email to pre-fxa subscribers
