+++ db/DB.js
@@ -79,6 +79,18 @@ const DB = {
     return subscriberAndEmails;
   },
 
+  async getPreFxaSubscribers() {
+    // We don't want to get subscribers records where the email address has
+    // since been added "under" an FxA subscription
+    const allEmailAddressesSubquery = knex("email_addresses").select("email");
+    const preFxaSubscribers = await knex("subscribers").where({
+      "fxa_uid": "",
+    }).andWhere(
+      "primary_email", "not in", allEmailAddressesSubquery
+    );
+    return preFxaSubscribers;
+  },
+
   async getSubscriberByEmail(email) {
     const [subscriber] = await knex("subscribers").where({
       "primary_email": email,
