+++ locale-utils.js
@@ -61,6 +61,7 @@ const LocaleUtils = {
     }
     log.info("LocaleUtils.init", {availableLanguages});
     log.info("LocaleUtils.init", {fluentBundles});
+    return {availableLanguages, fluentBundles};
   },
 
   loadLanguagesIntoApp (app) {
