+++ routes/hibp.js
@@ -6,7 +6,7 @@ const bodyParser = require("body-parser");
 const bearerToken = require("express-bearer-token");
 
 const {asyncMiddleware} = require("../middleware");
-const {notify, breaches, sendEmailToPreFxaSubscribers} = require("../controllers/hibp");
+const {notify, breaches} = require("../controllers/hibp");
 
 
 const router = express.Router();
@@ -15,6 +15,5 @@ const jsonParser = bodyParser.json();
 router.use("/notify", bearerToken());
 router.post("/notify", jsonParser, asyncMiddleware(notify));
 router.get("/breaches", jsonParser, asyncMiddleware(breaches));
-router.get("/sendPreFxaEmail", jsonParser, asyncMiddleware(sendEmailToPreFxaSubscribers));
 
 module.exports = router;
